Refactor: remove repository
