+++ src/main/java/glue/Gachi_Sanchaek/walk/service/WalkService.java
@@ -15,7 +15,6 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 @RequiredArgsConstructor
 public class WalkService {
-    private final WalkRecordRepository walkRecordRepository;
     private final WalkRecordService walkRecordService;
     private final RewardService rewardService;
     private final VerificationService verificationService;
@@ -31,16 +30,6 @@ public class WalkService {
         return rewardService.finalizeWalk(userId,walk,"산책 종료 완료");
     }
 
-    //웹소켓 연결 시 status = ONGOING (QR인증 제외)
-    public void onWebSocketConnect(Long walkId) {
-        WalkRecord walk = walkRecordRepository.findById(walkId)
-                .orElseThrow(()->new IllegalArgumentException("해당 산책 세션이 존재하지 않습니다."));
-        if(walk.getVerificationMethod()!= VerificationMethod.QR
-                && walk.getStatus() == WalkStatus.WAITING){
-            walk.setStatus(WalkStatus.ONGOING);
-            walkRecordRepository.save(walk);
-        }
-    }
 
     public Object handleQrScan(Long userId, String qrToken){
         return verificationService.handleQrScan(userId,qrToken);
